+++ tests/acceptance/workload_identity_test.go
@@ -2,13 +2,17 @@ package acceptance_test
 
 import (
 	"context"
+	"fmt"
+	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	serviceaccount "github.com/giantswarm/workload-identity-operator-gcp/controllers"
 	"github.com/giantswarm/workload-identity-operator-gcp/webhook"
 )
 
@@ -16,32 +20,31 @@ var _ = Describe("Workload Identity", func() {
 	var (
 		ctx context.Context
 		pod *corev1.Pod
+
+		serviceAccount *corev1.ServiceAccount
+
+		gcpServiceAccount    = "service-account@email"
+		workloadIdentityPool = "workload-identity-pool-id"
+		identityProvider     = "https://gkehub.googleapis.com/projects/testing-1234/locations/global/memberships/cluster"
+
+		timeout  = time.Second * 5
+		interval = time.Millisecond * 250
 	)
 
 	BeforeEach(func() {
 		ctx = context.Background()
-		serviceAccount := &corev1.ServiceAccount{
+		serviceAccount = &corev1.ServiceAccount{
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      "the-service-account",
 				Namespace: namespace,
 				Annotations: map[string]string{
-					webhook.AnnotationGCPServiceAccount:      "service-account@email",
-					webhook.AnnotationWorkloadIdentityPoolID: "workload-identity-pool-id",
+					webhook.AnnotationGCPServiceAccount:      gcpServiceAccount,
+					webhook.AnnotationWorkloadIdentityPoolID: workloadIdentityPool,
+					webhook.AnnotationGCPIdentityProvider:    identityProvider,
 				},
 			},
 		}
 		Expect(k8sClient.Create(ctx, serviceAccount)).To(Succeed())
-
-		configMap := &corev1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "the-service-account-google-application-credentials",
-				Namespace: namespace,
-			},
-			Data: map[string]string{
-				"config": "the-content",
-			},
-		}
-		Expect(k8sClient.Create(ctx, configMap)).To(Succeed())
 
 		pod = &corev1.Pod{
 			ObjectMeta: metav1.ObjectMeta{
@@ -74,6 +77,41 @@ var _ = Describe("Workload Identity", func() {
 		}
 	})
 
+	It("Creates the secret with the credentials needed", func() {
+		secret := &corev1.Secret{}
+		secretName := fmt.Sprintf("%s-%s", serviceAccount.Name, serviceaccount.SecretNameSuffix)
+
+		Eventually(func() error {
+			err := k8sClient.Get(ctx, client.ObjectKey{
+				Namespace: namespace,
+				Name:      secretName,
+			}, secret)
+
+			return err
+
+		}, timeout, interval).Should(BeNil())
+
+		Expect(secret).ToNot(BeNil())
+		Expect(secret.Name).To(Equal(secretName))
+		Expect(secret.Namespace).To(Equal(namespace))
+
+		data := string(secret.Data["config"])
+
+		expectedData := fmt.Sprintf(`{
+	     "type": "external_account",
+	     "audience": "identitynamespace:%[1]s:%[2]s",
+	     "service_account_impersonation_url": "https://iamcredentials.googleapis.com/v1/projects/-/serviceAccounts/%[3]s:generateAccessToken",
+	     "subject_token_type": "urn:ietf:params:oauth:token-type:jwt",
+	     "token_url": "https://sts.googleapis.com/v1/token",
+	     "credential_source": {
+	       "file": "/var/run/secrets/tokens/gcp-ksa/token"
+	     }
+	   }`, workloadIdentityPool, identityProvider, gcpServiceAccount)
+
+		Expect(data).Should(MatchJSON(expectedData))
+
+	})
+
 	It("Injects the credentials file in the pod", func() {
 		Expect(k8sClient.Create(ctx, pod)).To(Succeed())
 
@@ -92,7 +130,7 @@ var _ = Describe("Workload Identity", func() {
 			return workload.Status.ContainerStatuses[0].Ready
 		}
 
-		Eventually(getPodStatus, "30s").Should(BeTrue(), "pod container failed to become ready")
+		Eventually(getPodStatus, "60s").Should(BeTrue(), "pod container failed to become ready")
 		Consistently(getPodStatus, "5s").Should(BeTrue(), "pod container errored")
 	})
 })